+++ LogicaNegocio/Entidades/Empleado.cs
@@ -30,7 +30,7 @@ namespace LogicaNegocio.Entidades
         public TipoEmpleado TipoEmpleado { get; set; }
         public string? CuentaBanco { get; set; }
         public string? Banco { get; set; }
-        public double IncentivoXHora { get; set; }
+        public decimal IncentivoXHora { get; set; }
 
         public bool Activo {  get; set; }
